Refactor registries
